[AMD] Implemented hipError discard after a failure to get pointer attributes (#7832)
Since ROCm7.x the error behavior was changed to match its' CUDA
counterpart. After unsuccessful getPointerAttr call the error was not
discarded and could lead to breaking the Pytorch state when allocating
memory after such call.

This PR is created to fix a gating issue and we would like it to merged
asap, but there can probably be more places in the code which lead to
same results. I didn't conditionalize it for rocm7.0 because for earlier
version hipGetLastError would just return nothing of value and get
discarded anyway without breaking any logic.